requirements.txt
streamlit==1.32.0
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
pytesseract==0.3.10
pyzbar==0.1.9
pillow==10.0.1
pandas==2.1.3
openpyxl==3.1.2
opencv-python==4.8.1.78
SQLAlchemy
Flask-SQLAlchemy
Werkzeug
python-dotenv
Flask-Migrate
xlsxwriter==3.2.5